python/requirements.txt
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
six==1.16.0
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
python-dateutil==2.8.2
